+++ assets/javascripts/kitten/components/grid/adaptable-grid/index.js
@@ -107,7 +107,6 @@ AdaptableGrid.defaultProps = {
 }
 
 const StyledGrid = styled.div`
-  width: 100%;
   box-sizing: border-box;
   display: flex;
   flex-wrap: wrap;