+++ assets/javascripts/kitten/components/date-picker/standalone-range/index.js
@@ -193,40 +193,23 @@ const StyledDatePicker = styled.div`
 
 export class StandaloneRangeDatePicker extends PureComponent {
   static propTypes = {
-    numberOfMonths: PropTypes.number,
-    locale: PropTypes.string,
-    previousMonth: PropTypes.string,
-    weekDays: PropTypes.array,
-    months: PropTypes.array,
-    navbarElement: PropTypes.node,
-    disabledDays: PropTypes.array,
-    title: PropTypes.string,
-    dayPickerProps: PropTypes.shape({}),
     styles: PropTypes.object,
-    firstDayOfWeek: PropTypes.number,
+    months: PropTypes.array,
+    previousMonth: PropTypes.string,
+    nextMonth: PropTypes.string,
   }
 
   static defaultProps = {
-    numberOfMonths: 2,
-    locale: 'en',
-    previousMonth: 'Previous month',
-    nextMonth: 'Next month',
-    weekDays: null,
-    title: '',
-    months: null,
-    navbarElement: '',
-    firstDayOfWeek: 1,
-    disabledDays: [
-      {
-        after: new Date(),
-      },
-    ],
-    dayPickerProps: {},
-
+    months: '',
+    previousMonth: '',
+    nextMonth: '',
     styles: {
       header: {
         backgroundColor: COLORS.background1,
         color: COLORS.font1,
+        icon: {
+          color: COLORS.font1,
+        },
       },
       borderColor: COLORS.line1,
       weekdaysColor: COLORS.font1,
@@ -289,6 +272,7 @@ export class StandaloneRangeDatePicker extends PureComponent {
       title,
       firstDayOfWeek,
       initialMonth,
+      numberOfMonths,
       ...datePickerProps
     } = this.props
 
@@ -299,7 +283,7 @@ export class StandaloneRangeDatePicker extends PureComponent {
       <StyledDatePicker styles={styles}>
         <DayPicker
           className="Selectable"
-          numberOfMonths={this.props.numberOfMonths}
+          numberOfMonths={numberOfMonths}
           selectedDays={[from, { from, to }]}
           modifiers={modifiers}
           onDayClick={this.handleDayClick}