+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Remove `defaultProps` in `StandaloneRangeDatePicker` component.
+- Fix: Fix `AdaptableGrid` width.
 
 ## [2.22.0] - 2019-08-28