+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Update `TextInputWithIcon` with `styled-components`.
+- Feature: Add `TextInputWithIcon` component.
 
 ## [2.23.0] - 2019-08-30