+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Export `ICON_TINY` and `ICON_BIG` css from `Button`.
+
 ## [2.23.0] - 2019-08-30
 
 Feature: